+++ routes/requestHandlers.js
@@ -1,14 +1,5 @@
-var exec = require("child_process").exec;
-var database = require("./database");
-
-function index(response, postData) {
-    console.log("[INFO] Request handler 'index' was called.");
-    exec("ls -lah", function (error, stdout, stderr) {
-        response.writeHead(200, {"Content-Type": "text/plain"});
-        response.write(stdout);
-        response.end();
-    });
-}
+var database = require("../modules/database");
+var responses = require("../modules/responses");
 
 function vote(response, postData) {
     //Test this endpoint with curl -d '{"vote": "a"}' -H "Content-Type: application/json" http://localhost:8000/vote
@@ -20,7 +11,7 @@ function vote(response, postData) {
         err["httpStatus"] = 400;
         err["httpResponse"] = "400 Bad Request";
         err["friendlyName"] = "JSON parse error";
-        errorResponse(err, response);
+        responses.errorResponse(err, response);
         return;
     }
 
@@ -33,39 +24,16 @@ function vote(response, postData) {
             if (!err["friendlyName"]) {
                 err["friendlyName"] = "Error recording vote";
             }
-            errorResponse(err, response);
+            responses.errorResponse(err, response);
             return;
         }
         else {
             console.log("[INFO] Logged vote to database");
-            successResponse(response);
+            responses.successResponse(response);
             return;
         }
     });
 }
 
-
-function errorResponse(err, response) {
-    message = {'status':'error', 'message':err['friendlyName']};
-
-    response.writeHead(err["httpStatus"], {"Content-Type": "application/json"});
-    response.write(err["httpResponse"] + "\n");
-    response.write(JSON.stringify(message) + "\n");
-    response.end();
-
-    console.log("[ERROR]", err)
-}
-
-function successResponse(response) {
-    message = {'status':'success'};
-
-    response.writeHead(200, {"Content-Type": "application/json"});
-    response.write("200 OK" + "\n");
-    response.write(JSON.stringify(message) + "\n");
-    response.end();
-}
-
-
-exports.index = index;
 exports.vote = vote;
